[BE] refactor: 냄새나는 코드 리팩터링

+++ backend/business-service/src/main/java/com/footprints/businessservice/app/domain/board/article/service/ArticleServiceImpl.java
@@ -42,6 +42,8 @@ public class ArticleServiceImpl implements ArticleService {
     private final TransferRepository transferRepository;
     private final ImageService imageService;
 
+    private static final String TRANSFER = "transfer";
+
     @Override
     public List<ArticleDto> getArticleList(SortCondition condition, Pageable pageable) {
         PageRequest pageRequest = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(),
@@ -49,11 +51,9 @@ public class ArticleServiceImpl implements ArticleService {
 
         Page<Article> articles = articleRepository.getArticleList(condition, pageRequest);
 
-        List<ArticleDto> result = articles.stream()
-                .map(article -> new ArticleDto(article))
+        return articles.stream()
+                .map(ArticleDto::new)
                 .collect(Collectors.toList());
-
-        return result;
     }
 
     @Override
@@ -70,7 +70,7 @@ public class ArticleServiceImpl implements ArticleService {
                 .category(request.getArticleRequest().getCategory())
                 .build();
 
-        if (request.getArticleRequest().getCategory().equals("transfer")) {
+        if (request.getArticleRequest().getCategory().equals(TRANSFER)) {
             transferRepository.save(
                     Transfer.builder()
                             .roomType(request.getTransferRequest().getRoomType())
@@ -99,7 +99,7 @@ public class ArticleServiceImpl implements ArticleService {
 
         List<Comment> findComments = articleRepository.getCommentList(articleId);
         List<CommentDto> comments = findComments.stream()
-                .map(comment -> new CommentDto(comment))
+                .map(CommentDto::new)
                 .collect(Collectors.toList());
 
         List<ImageDto> images = article.getImages().stream()
@@ -110,7 +110,7 @@ public class ArticleServiceImpl implements ArticleService {
                         .build())
                 .collect(Collectors.toList());
 
-        if (article.getCategory().equals("transfer")) {
+        if (article.getCategory().equals(TRANSFER)) {
             Transfer transfer = transferRepository.getTransferByArticleId(articleId);
             TransferDto transferDto = transfer.toDto(transfer);
 
@@ -157,11 +157,9 @@ public class ArticleServiceImpl implements ArticleService {
     public List<ArticleDto> searchArticle(SearchCondition condition, Pageable pageable) {
         Page<Article> articles = articleRepository.searchArticle(condition, pageable);
 
-        List<ArticleDto> result = articles.stream()
-                .map(article -> new ArticleDto(article))
+        return articles.stream()
+                .map(ArticleDto::new)
                 .collect(Collectors.toList());
-
-        return result;
     }
 
     @Override
@@ -201,11 +199,9 @@ public class ArticleServiceImpl implements ArticleService {
 
         Page<ScrappedArticle> scrappedArticles = scrappedArticleRepository.getScrappedArticleList(Long.parseLong(memberId), category, pageRequest);
 
-        List<ScrappedArticleDto> result = scrappedArticles.stream()
-                .map(scrappedArticle -> new ScrappedArticleDto(scrappedArticle))
+        return scrappedArticles.stream()
+                .map(ScrappedArticleDto::new)
                 .collect(Collectors.toList());
-
-        return result;
     }
 
     @Override
@@ -222,8 +218,12 @@ public class ArticleServiceImpl implements ArticleService {
     @Override
     @Transactional
     public void deleteArticle(String memberId, Long articleId) {
-        Article article = articleRepository.getArticle(articleId);
-        transferRepository.delete(transferRepository.getTransferByArticleId(articleId));
+        Article article = articleRepository.getArticleWithCommentList(articleId);
+
+        if (article.getCategory().equals(TRANSFER)) {
+            transferRepository.delete(transferRepository.getTransferByArticleId(articleId));
+        }
+
         articleRepository.delete(article);
     }
